+++ program/Exponentiation.h
@@ -1,13 +1,14 @@
 #ifndef EXPONENTIATION_H
 #define EXPONENTIATION_H
 
+#include <math.h>
 #include "BinaryOperator.h"
 
 class Exponentiation : public BinaryOperator {
-    public:
-        float Calculate(float lhs, float rhs) {
-            return pow(lhs, rhs);
-        };
+public:
+	double Calculate(double lhs, double rhs) {
+		return pow(lhs, rhs);
+	};
 };
 
-#endif
+#endif // EXPONENTIATION_H
